Moved Controls into its own component
